Add new scripts for SQLite database management and schema verification
- Introduced `check_sqlite.py` to display database configuration and table information.
- Added `check_table.py` for schema inspection of the `historial_mora_detalle` table.
- Created `debug_model.py` to generate and verify SQLAlchemy DDL for `HistorialMoraDetalle`.
- Implemented `fix_ahsaldia_patron.py` to update the AHSALDIA pattern in parameters.
- Developed `fix_pk_integer.py` to correct primary key types from BIGINT to INTEGER for autoincrement functionality.
- Added `fix_schema.py` to verify and correct the schema of the `logs_cp` table.
- Created `inspect_headers.py` to normalize headers from .lis files.
- Added `params_prueba.py` for testing parameter updates in the database.
- Implemented `test_lector_cadetacaco.py` for quick testing of the cadetacaco reader against real files.

These scripts enhance database management, schema verification, and testing capabilities within the preventive management system.

+++ src/preventiva/infrastructure/persistence/models/ejecucion_pad.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Optional
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String
+from sqlalchemy import DateTime, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from preventiva.infrastructure.persistence.base import Base
@@ -10,7 +10,7 @@ from preventiva.infrastructure.persistence.base import Base
 class EjecucionPad(Base):
     __tablename__ = "ejecucion_pad"
 
-    id:              Mapped[int]           = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id:              Mapped[int]           = mapped_column(Integer, primary_key=True, autoincrement=True)
     proceso_cod:     Mapped[str]           = mapped_column(String(14), ForeignKey("historial_proceso.proceso_cod"), nullable=False)
     paso_ejecucion:  Mapped[str]           = mapped_column(String(100), nullable=False)
     estado:          Mapped[str]           = mapped_column(String(10),  nullable=False)
